ENGR00274713 Adjust makefile to support patchless yocto build
Yocto build is different with other cross build. In the previous
yocto release, a special patch was created by yocto developpers. The
patch highly depends on code status and can't be reused for next
release. It is better to refine the makefile so yocto developpers
need not create patch any more.

Date: 2013-08-12
Signed-off-by: Zhenyong Chen <b07273@freescale.com>